Refactor component imports and update copyright year
